test: asset manager desired state tests.

+++ agent/test/asset-manager.test.js
@@ -54,20 +54,17 @@ test.afterEach.always('cleanup', async t => {
 })
 
 function getEmptyDeviceState(states) {
-  const fileName = 'asset_1.json'
-  const _states = states.map(state => {
-  switch (state.type) {
-    case 'desired':
-    case 'reported':
-      return {
-        type: state.type,
-        state: {}
-      }
-    case 'status':
-      return Utils.getDummyStatusState("enebular-agent", agentVer)
-    }
-  })
-  return _states
+  return [
+    {
+      type: 'desired',
+      state: {}
+    },
+    {
+      type: 'reported',
+      state: {}
+    },
+    Utils.getDummyStatusState("enebular-agent", agentVer)
+  ]
 }
 
 function addAsset(desiredState, assetId, fileName, integrity) {
@@ -96,10 +93,15 @@ function delAsset(desiredState, assetId) {
   return Utils.getDummyState('desired', desiredState.state)
 }
 
+function modifyAsset(desiredState, assetId, prop, value) {
+  objectPath.set(desiredState,
+      'state.assets.assets.' + assetId + '.' + prop, value)
+  return Utils.getDummyState('desired', desiredState.state)
+}
+
 async function initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath, 
     updateReq, populateDesiredState) {
   server.onDeviceStateGet = (req, res) => {
-    deviceStates = getEmptyDeviceState(req.body.states)
     populateDesiredState(deviceStates[0])
     res.send({ states: deviceStates })
   }
@@ -125,34 +127,68 @@ async function initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath,
     DummyServerPort
   )
   agent = ret.agent
+  return ret
+}
 
+async function createAgentWithAssetsDeployed(t, assetCount, cleanup) {
+  let tmpAssetDataPath = '/tmp/tmp-asset-data-' + Utils.randomString()
+  let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
+  let deviceStates = getEmptyDeviceState(['desired', 'reported', 'status'])
+  let reportedStates = []
+  let assets = []
+  let randomDataPath = path.join(__dirname, 'data', 'tmp')
+  let assetsCount = assetCount
+
+  if (!fs.existsSync(randomDataPath))
+    fs.mkdirSync(randomDataPath)
+
+  for (let i = 0; i < assetsCount; i++) {
+    let id = 'random-' + Utils.randomString()
+    let p = path.join(__dirname, 'data', 'tmp', id)
+    await Utils.createFileOfSize(p, 1024 * 10)
+    const integrity = await Utils.getFileIntegrity(p)
+    assets.push({
+      id: id,
+      name: id,
+      integrity: integrity
+    })
+  }
+  let ret = await initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath, reportedStates, (desiredState) => {
+    assets.map((asset) => {
+      addAsset(desiredState, asset.id, asset.name, asset.integrity)
+    })
+  })
+
+  await polling(() => { 
+      return !agent._assetManager._getFirstPendingChangeAsset() && !agent._assetManager._processingChanges
+  }, 0, 1000, assetsCount * 2000)
+
+  const cacheStates = JSON.parse(fs.readFileSync(tmpAssetStatePath, 'utf8'))
+  cacheStates.map((state, index) => {
+    t.is(state.id, assets[index].id)
+    t.is(state.state, 'deployed')
+    t.true(fs.existsSync(tmpAssetDataPath + '/dst/' + assets[index].name))
+  })
+
+  if (cleanup) {
+    fs.unlinkSync(tmpAssetStatePath)
+    fs.removeSync(tmpAssetDataPath)
+    fs.removeSync(randomDataPath)
+  }
+
+  return {
+    assetStatePath: tmpAssetStatePath,
+    assetDataPath: tmpAssetDataPath,
+    randomDataPath: randomDataPath,
+    deviceStates: deviceStates,
+    reportedStates: reportedStates,
+    assets: assets,
+    connector: ret.connector
+  }
 }
 
 test.serial('AssetManager.1: Agent deploys asset according to desired state', async t => {
-  let tmpAssetDataPath = '/tmp/tmp-asset-data-' + Utils.randomString()
-  let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
-  let deviceStates
-  let updateReq = []
-  let newAssetId = Utils.randomString()
-  let assetName = 'asset_1.json'
-  const integrity = await Utils.getFileIntegrity(path.join(__dirname, 'data', assetName))
-
-  await initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath, updateReq, (desiredState) => {
-    addAsset(desiredState, newAssetId, assetName, integrity)
-  })
-
-  await polling(() => { return updateReq.length > 0 }, 0, 100, 10000)
-
-  t.is(updateReq[0].type, 'reported')
-  t.is(updateReq[0].op, 'set')
-  t.is(updateReq[0].state.state, 'deployed')
-
-  const state = JSON.parse(fs.readFileSync(tmpAssetStatePath, 'utf8'))
-  t.is(state[0].id, newAssetId)
-  t.is(state[0].state, 'deployed')
-  fs.unlinkSync(tmpAssetStatePath)
-  t.true(fs.existsSync(tmpAssetDataPath + '/dst/asset_1.json'))
-  fs.removeSync(tmpAssetDataPath)
+  await createAgentWithAssetsDeployed(t, 1, true)
 })
 
 test.serial('AssetManager.2: Agent handles asset deploy failure(file integrity)', async t => {
@@ -160,7 +196,7 @@ test.serial('AssetManager.2: Agent handles asset deploy failure(file integrity)'
   let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
   let newAssetId = Utils.randomString()
   let updateReq = []
-  let deviceStates
+  let deviceStates = getEmptyDeviceState(['desired', 'reported', 'status'])
   let assetName = 'asset_1.json'
 
   await initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath, updateReq, (desiredState) => {
@@ -190,7 +226,7 @@ test.serial('AssetManager.3: Agent handles asset deploy failure(download file)',
   let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
   let newAssetId = Utils.randomString()
   let updateReq = []
-  let deviceStates
+  let deviceStates = getEmptyDeviceState(['desired', 'reported', 'status'])
   let assetName = 'asset_1.json'
   const integrity = await Utils.getFileIntegrity(path.join(__dirname, 'data', assetName))
 
@@ -217,81 +253,19 @@ test.serial('AssetManager.3: Agent handles asset deploy failure(download file)',
 })
 
 test.serial('AssetManager.4: Agent deploys multiple assets according to desired state', async t => {
-  let tmpAssetDataPath = '/tmp/tmp-asset-data-' + Utils.randomString()
-  let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
-  let deviceStates
-  let updateReq = []
-  let assets = []
-  let ramdomDataPath = path.join(__dirname, 'data', 'tmp')
-  let assetsCount = 20
-
-  if (!fs.existsSync(ramdomDataPath))
-    fs.mkdirSync(ramdomDataPath)
-
-  for (let i = 0; i < assetsCount; i++) {
-    let id = 'random-' + Utils.randomString()
-    let p = path.join(__dirname, 'data', 'tmp', id)
-    await Utils.createFileOfSize(p, 1024 * 10)
-    const integrity = await Utils.getFileIntegrity(p)
-    assets.push({
-      id: id,
-      name: id,
-      integrity: integrity
-    })
-  }
-  await initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath, updateReq, (desiredState) => {
-    assets.map((asset) => {
-      addAsset(desiredState, asset.id, asset.name, asset.integrity)
-    })
-  })
-
-  await polling(() => { 
-      return !agent._assetManager._getFirstPendingChangeAsset()
-  }, 0, 1000, assetsCount * 1500)
-
-  const cacheStates = JSON.parse(fs.readFileSync(tmpAssetStatePath, 'utf8'))
-  cacheStates.map((state, index) => {
-    t.is(state.id, assets[index].id)
-    t.is(state.state, 'deployed')
-    t.true(fs.existsSync(tmpAssetDataPath + '/dst/' + assets[index].name))
-  })
-
-  fs.unlinkSync(tmpAssetStatePath)
-  fs.removeSync(tmpAssetDataPath)
-  fs.removeSync(ramdomDataPath)
+  await createAgentWithAssetsDeployed(t, 20, true)
 })
 
-test.serial('AssetManager.5: Agent deploys asset via desired state', async t => {
+test.serial('AssetManager.5: Agent deploys new asset via deviceStateChange', async t => {
   let tmpAssetDataPath = '/tmp/tmp-asset-data-' + Utils.randomString()
   let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
-  let deviceStates
+  let deviceStates = getEmptyDeviceState(['desired', 'reported', 'status'])
   let updateReq = []
-  server.onDeviceStateGet = (req, res) => {
-    deviceStates = getEmptyDeviceState(req.body.states)
-    res.send({ states: deviceStates })
-  }
 
-  server.onDeviceStateUpdate = (req, res) => {
-    const result = req.body.updates.map(update => {
-      updateReq.push(update)
-      return {
-        success: true,
-        meta: {}
-      }
-    })
-    res.send({ updates: result })
-  }
+  let ret = await initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath,
+      updateReq, (desiredState) => {
+  })
 
-  const ret = await createAuthenticatedAgent(
-    t,
-    server,
-    Utils.addNodeRedPortToConfig({
-      ENEBULAR_ASSETS_DATA_PATH: tmpAssetDataPath,
-      ENEBULAR_ASSETS_STATE_PATH: tmpAssetStatePath
-    }, NodeRedPort),
-    DummyServerPort
-  )
-  agent = ret.agent
   t.true(await nodeRedIsAlive(NodeRedPort))
 
   t.is(updateReq.length, 0)
@@ -300,7 +274,8 @@ test.serial('AssetManager.5: Agent deploys asset via desired state', async t => 
   let assetName = 'asset_1.json'
   const integrity = await Utils.getFileIntegrity(path.join(__dirname, 'data', assetName))
 
-  let desiredState = await addAsset(deviceStates[0], newAssetId, assetName, integrity)
+  // Deplay new asset
+  let desiredState = addAsset(deviceStates[0], newAssetId, assetName, integrity)
   ret.connector.sendMessage('deviceStateChange', {
     type: 'desired',
     op: 'set',
@@ -309,11 +284,223 @@ test.serial('AssetManager.5: Agent deploys asset via desired state', async t => 
     state: desiredState.state.assets.assets[newAssetId]
   })
 
-  await polling(() => { return updateReq.length > 0 }, 0, 100, 5000)
+  await polling(() => { 
+      return !agent._assetManager._getFirstPendingChangeAsset()
+  }, 2000, 1000, 5000)
 
-  fs.removeSync(tmpAssetDataPath)
+  const cacheStates = JSON.parse(fs.readFileSync(tmpAssetStatePath, 'utf8'))
+
+  t.is(cacheStates[0].id, newAssetId)
+  t.is(cacheStates[0].state, 'deployed')
   fs.unlinkSync(tmpAssetStatePath)
+  t.true(fs.existsSync(tmpAssetDataPath + '/dst/asset_1.json'))
+  fs.removeSync(tmpAssetDataPath)
 })
+
+test.serial('AssetManager.6: Agent re-deploys existing asset via deviceStateChange', async t => {
+  let tmpAssetDataPath = '/tmp/tmp-asset-data-' + Utils.randomString()
+  let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
+  let deviceStates = getEmptyDeviceState(['desired', 'reported', 'status'])
+  let updateReq = []
+  let newAssetId = Utils.randomString()
+  let assetName = 'asset_1.json'
+  const integrity = await Utils.getFileIntegrity(path.join(__dirname, 'data', assetName))
+
+  let ret = await initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath, updateReq, (desiredState) => {
+    addAsset(desiredState, newAssetId, assetName, integrity)
+  })
+
+  await polling(() => { return updateReq.length > 0 }, 0, 100, 10000)
+
+  t.is(updateReq[0].type, 'reported')
+  t.is(updateReq[0].op, 'set')
+  t.is(updateReq[0].state.state, 'deployed')
+
+  // Deplay again
+  let updateId = Utils.randomString()
+  let desiredState = modifyAsset(deviceStates[0], newAssetId, 'updateId', updateId)
+  ret.connector.sendMessage('deviceStateChange', {
+    type: 'desired',
+    op: 'set',
+    path: 'assets.assets.' + newAssetId,
+    meta: desiredState.meta,
+    state: desiredState.state.assets.assets[newAssetId]
+  })
+
+  await polling(() => { 
+      return !agent._assetManager._getFirstPendingChangeAsset()
+  }, 2000, 1000, 5000)
+
+  t.is(updateReq[1].state.state, 'deployPending')
+  t.is(updateReq[2].state.state, 'removing')
+  t.is(updateReq[3].state.state, 'deploying')
+  t.is(updateReq[4].state.state, 'deployed')
+  t.true(fs.existsSync(tmpAssetDataPath + '/dst/asset_1.json'))
+  const state = JSON.parse(fs.readFileSync(tmpAssetStatePath, 'utf8'))
+  t.is(state[0].id, newAssetId)
+  t.is(state[0].state, 'deployed')
+  t.is(state[0].updateId, updateId)
+  t.true(fs.existsSync(tmpAssetDataPath + '/dst/asset_1.json'))
+
+  fs.unlinkSync(tmpAssetStatePath)
+  fs.removeSync(tmpAssetDataPath)
+})
+
+test.serial('AssetManager.7: Agent removes existing asset via deviceStateChange', async t => {
+  let tmpAssetDataPath = '/tmp/tmp-asset-data-' + Utils.randomString()
+  let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
+  let deviceStates = getEmptyDeviceState(['desired', 'reported', 'status'])
+  let updateReq = []
+  let newAssetId = Utils.randomString()
+  let assetName = 'asset_1.json'
+  const integrity = await Utils.getFileIntegrity(path.join(__dirname, 'data', assetName))
+
+  let ret = await initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath, updateReq, (desiredState) => {
+    addAsset(desiredState, newAssetId, assetName, integrity)
+  })
+
+  await polling(() => { return updateReq.length > 0 }, 0, 100, 10000)
+
+  t.is(updateReq[0].type, 'reported')
+  t.is(updateReq[0].op, 'set')
+  t.is(updateReq[0].state.state, 'deployed')
+
+  let updateId = Utils.randomString()
+  let desiredState = delAsset(deviceStates[0], newAssetId)
+  ret.connector.sendMessage('deviceStateChange', {
+    type: 'desired',
+    op: 'remove',
+    path: 'assets.assets.' + newAssetId,
+    meta: desiredState.meta,
+    state: desiredState.state.assets.assets[newAssetId]
+  })
+
+  await polling(() => { 
+      return !agent._assetManager._getFirstPendingChangeAsset()
+  }, 2000, 1000, 5000)
+
+  console.log(updateReq)
+
+  t.is(updateReq[1].state.state, 'removePending')
+  t.is(updateReq[2].state.state, 'removing')
+  t.is(updateReq[3].op, 'remove')
+  t.false(fs.existsSync(tmpAssetDataPath + '/dst/asset_1.json'))
+  const state = JSON.parse(fs.readFileSync(tmpAssetStatePath, 'utf8'))
+  t.true(state.length == 0)
+
+  fs.unlinkSync(tmpAssetStatePath)
+  fs.removeSync(tmpAssetDataPath)
+})
+
+test.serial('AssetManager.8: Agent handles removing asset failure', async t => {
+  let tmpAssetDataPath = '/tmp/tmp-asset-data-' + Utils.randomString()
+  let tmpAssetStatePath = '/tmp/tmp-asset-state-' + Utils.randomString()
+  let deviceStates = getEmptyDeviceState(['desired', 'reported', 'status'])
+  let updateReq = []
+  let newAssetId = Utils.randomString()
+  let assetName = 'asset_1.json'
+  const integrity = await Utils.getFileIntegrity(path.join(__dirname, 'data', assetName))
+
+  let ret = await initAgent(t, deviceStates, tmpAssetDataPath, tmpAssetStatePath, updateReq, (desiredState) => {
+    addAsset(desiredState, newAssetId, assetName, integrity)
+    // addAsset(desiredState, Utils.randomString(), 'asset_2.json', integrity)
+  })
+
+  await polling(() => { 
+      return !agent._assetManager._getFirstPendingChangeAsset()
+  }, 2000, 1000, 5000)
+
+  t.is(updateReq[0].type, 'reported')
+  t.is(updateReq[0].op, 'set')
+  t.is(updateReq[0].state.state, 'deployed')
+
+  // change file to directory which fails the removing
+  fs.removeSync(tmpAssetDataPath + '/dst/asset_1.json')
+  fs.mkdirSync(tmpAssetDataPath + '/dst/asset_1.json')
+
+  let desiredState = delAsset(deviceStates[0], newAssetId)
+  ret.connector.sendMessage('deviceStateChange', {
+    type: 'desired',
+    op: 'remove',
+    path: 'assets.assets.' + newAssetId,
+    meta: desiredState.meta,
+    state: desiredState.state.assets.assets[newAssetId]
+  })
+
+  await polling(() => { 
+      return !agent._assetManager._getFirstPendingChangeAsset()
+  }, 2000, 1000, 5000)
+
+  t.is(updateReq[1].state.state, 'removePending')
+  t.is(updateReq[2].state.state, 'removing')
+  t.is(updateReq[3].state.state, 'removeFail')
+  t.true(fs.existsSync(tmpAssetDataPath + '/dst/asset_1.json'))
+  const state = JSON.parse(fs.readFileSync(tmpAssetStatePath, 'utf8'))
+
+  // console.log(JSON.stringify(state, null, 2))
+
+  t.is(state[0].id, newAssetId)
+  t.is(state[0].state, 'removeFail')
+
+  fs.unlinkSync(tmpAssetStatePath)
+  fs.removeSync(tmpAssetDataPath)
+})
+
+test.serial('AssetManager.9: Agent handles multiple assets changes in desired status', async t => {
+  let ret = await createAgentWithAssetsDeployed(t, 3, false)
+  let newAssetId = Utils.randomString()
+  let assetName = 'asset_1.json'
+  const integrity = await Utils.getFileIntegrity(path.join(__dirname, 'data', assetName))
+
+  ret.reportedStates.length = 0
+  // remove
+  let desiredState = delAsset(ret.deviceStates[0], ret.assets[0].id)
+  // modify
+  let updateId = Utils.randomString()
+  desiredState = modifyAsset(ret.deviceStates[0], ret.assets[1].id, 'updateId', updateId)
+  // add
+  desiredState = addAsset(ret.deviceStates[0], newAssetId, assetName, integrity)
+
+  // console.log(JSON.stringify(desiredState, null, 2))
+
+  server.onDeviceStateGet = (req, res) => {
+    ret.deviceStates[0] = desiredState
+    res.send({ states: ret.deviceStates })
+  }
+
+  // send message without meta data will trigger a desired status refresh.
+  ret.connector.sendMessage('deviceStateChange', {
+    type: 'desired',
+    op: 'remove',
+    path: 'assets.assets.' + newAssetId,
+    state: desiredState.state.assets.assets[newAssetId]
+  })
+
+  await polling(() => { 
+      return !agent._assetManager._getFirstPendingChangeAsset()
+  }, 2000, 1000, 5000)
+
+  // console.log(ret.reportedStates)
+  const state = JSON.parse(fs.readFileSync(ret.assetStatePath, 'utf8'))
+
+  t.is(state[0].id, ret.assets[1].id)
+  t.is(state[0].state, 'deployed')
+  t.is(state[0].updateId, updateId)
+
+  t.is(state[1].id, ret.assets[2].id)
+  t.is(state[1].state, 'deployed')
+
+  t.is(state[2].id, newAssetId)
+  t.is(state[2].state, 'deployed')
+
+  fs.unlinkSync(ret.assetStatePath)
+  fs.removeSync(ret.assetDataPath)
+  fs.removeSync(ret.randomDataPath)
+})
+
+
+
+
